FIX: problem with read access

+++ driver/flash.h
@@ -18,7 +18,7 @@
 #include <msp430.h>
 
 uint8_t flash_store(void *data, uint8_t size, uint16_t pos);
-uint8_t flash_read(void *data, uint8_t size, uint8_t pos);
+uint8_t flash_read(void *data, uint8_t size, uint16_t pos);
 void flash_erase(uint8_t *Flash_ptr);
 
 #endif /* FLASH_H_ */
